Adjusted import name in __main__.py
I feel dumb:

import .eggord

was working because it was a relative import to the package eggord, of
the eggord.py file in that package.

So the proper way to address that file is:

import eggord.eggord

Also removed some references to namepicker in config files that were
imported from that project.

Currently trying to get mypy to type check everything.