Got pylint working
The main issue is that pylint currently requires the isort version to be
<5, but I want the latest.

So I'm calling pylint from a pipx install, which means it doesn't have
access to the python interpreter that's installed in the virtual
environment.

pylint has an init-hook which is available specifically for munging
sys.path so it can pick up packages used in the maodule being checked.

Included an entire script in pyproject.toml to do this sys.path munging.

Took some work, but made it so the script in docs/pylint_init_hook.py is
this exact script, copy and pasted, so that the script can be manually
linted and type checked if changes need to be made.

Currently ideal configuration is to be able to run:

isort . && black . && mypy && pylint eggord tests

linting tests is a bit much for right now.