Version bumping working; tox py37 working
Version bumping happens solely through:
poetry version {major|minor|patch|etc}

Would like to add a test that calls poetry version to see if the tested
package version matches the package's __version__ string.

Also, tox setup currently has a hack where test requirements are
generated as their own "test", so that the file can be included in the
rest of the actual tests.

Also, added # NOTE:BUG markers around to indicate areas of less than
ideal code that are waiting on the resolution of open issues to be
improved.